Extract e2e dashboard helpers
